separate foreground check part from KcaViewButtonService
- check login success event for official app
