+++ app/src/main/java/com/antest1/kcanotify/KcaConstants.java
@@ -10,6 +10,8 @@ public final class KcaConstants {
     public final static String KC_WV_PACKAGE_NAME = "com.example.kcs_webview";
     public final static String DMMLOGIN_PACKAGE_NAME = "com.dmm.app.store";
     public final static String MAINACTIVITY_NAME = "com.antest1.kcanotify.MainActivity";
+    public final static String DMM_SDK_STORE_PACKAGE = "com.dmm.games.android.sdk.store";
+    public final static String KC_WV_START_ACTIVITY = "com.example.kcs_webview.activity.StartActivity";
     public final static String GOTO_PACKAGE_NAME = "com.antest1.gotobrowser";
 
     public static final String REQUEST_PATH = "/request";
