+++ app/src/main/java/com/antest1/kcanotify/KcaService.java
@@ -284,6 +284,11 @@ public class KcaService extends Service {
                     stopSelf();
                 }
 
+                if (getBooleanPreferences(getApplicationContext(), PREF_FAIRY_AUTOHIDE)) {
+                    startService(new Intent(this, KcaForegroundCheckService.class)
+                            .setAction(KcaForegroundCheckService.FAIRY_FORECHECK_ON));
+                }
+
                 KcaBattle.setHandler(nHandler);
                 KcaApiData.setHandler(nHandler);
                 KcaAlarmService.setHandler(nHandler);
@@ -347,6 +352,7 @@ public class KcaService extends Service {
                 stopService(new Intent(this, KcaViewButtonService.class));
                 stopService(new Intent(this, KcaExpeditionCheckViewService.class));
                 stopService(new Intent(this, KcaCustomToastService.class));
+                stopService(new Intent(this, KcaForegroundCheckService.class));
                 setServiceDown();
                 KcaAlarmService.clearAlarmCount();
                 stopSelfResult(startId);
